Factor process management scripts so "run" does not depend on "daemonize"
