+++ src/allmydata/scripts/tahoe_start.py
@@ -11,7 +11,7 @@ from allmydata.scripts.common import BasedirOptions
 from allmydata.scripts.default_nodedir import _default_nodedir
 from allmydata.util.encodingutil import quote_local_unicode_path
 
-from .tahoe_daemonize import MyTwistdConfig, identify_node_type
+from .run_common import MyTwistdConfig, identify_node_type
 
 
 class StartOptions(BasedirOptions):
